requirements.txt
# requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
scipy>=1.11.0
scikit-image>=0.21.0
opencv-contrib-python>=4.8.0
streamlit-drawable-canvas>=0.9.3
streamlit-webrtc>=0.63.11
opencv-python-headless>=4.12.0.88